Fix #402
