+++ src/Spaces/Fourier/Fourier.jl
@@ -54,9 +54,10 @@ hasfasttransform(::Hardy) = true
 typealias Taylor{D<:Domain} Hardy{true,D}
 
 plan_transform(::Taylor,x::Vector) = plan_fft(x)
-plan_itransform(::Taylor,x::Vector) = plan_ifft(x)
+plan_itransform{T<:Complex}(::Taylor,x::Vector{T}) = plan_ifft!(x) # we can reuse vector in itransform
+plan_itransform{T}(::Taylor,x::Vector{T}) = plan_ifft!(Array{Complex{T}}(length(x))) # we can reuse vector in itransform
 transform(::Taylor,vals::Vector,plan) = alternatesign!(plan*vals/length(vals))
-itransform(::Taylor,cfs::Vector,plan) = plan*alternatesign!(cfs)*length(cfs)
+itransform(::Taylor,cfs::Vector,plan) = plan*alternatesign!(cfs*length(cfs))
 
 plan_transform(::Hardy{false},x::Vector) = plan_fft(x)
 plan_itransform(::Hardy{false},x::Vector) = plan_ifft(x)
@@ -151,8 +152,8 @@ typealias Laurent{DD} SumSpace{Tuple{Hardy{true,DD},Hardy{false,DD}},ComplexBasi
 
 plan_transform{DD}(::Laurent{DD},x::Vector)=plan_svfft(x)
 plan_itransform{DD}(::Laurent{DD},x::Vector)=plan_isvfft(x)
-transform{DD}(::Laurent{DD},vals,plan)=svfft(vals,plan)
-itransform{DD}(::Laurent{DD},cfs,plan)=isvfft(cfs,plan)
+transform{DD}(::Laurent{DD},vals,plan...) = svfft(vals,plan...)
+itransform{DD}(::Laurent{DD},cfs,plan...) = isvfft(cfs,plan...)
 
 
 function evaluate{DD}(f::AbstractVector,S::Laurent{DD},z)
@@ -197,11 +198,18 @@ for T in (:CosSpace,:SinSpace)
 end
 
 points{D}(sp::Fourier{D},n)=points(domain(sp),n)
-plan_transform{T<:FFTW.fftwNumber,D}(::Fourier{D},x::Vector{T}) = FFTW.plan_r2r(x, FFTW.R2HC)
-plan_itransform{T<:FFTW.fftwNumber,D}(::Fourier{D},x::Vector{T}) = FFTW.plan_r2r(x, FFTW.HC2R)
+plan_transform{T<:FFTW.fftwNumber,D}(::Fourier{D},x::Vector{T}) =
+    FFTW.plan_r2r(x, FFTW.R2HC)
+plan_itransform{T<:FFTW.fftwNumber,D}(::Fourier{D},x::Vector{T}) =
+    FFTW.plan_r2r(x, FFTW.HC2R)
 
-plan_transform{D}(::Fourier{D},x::Vector)=error("transform for Fourier only implemented for fftwNumbers")
-plan_itransform{D}(::Fourier{D},x::Vector)=error("transform for Fourier only implemented for fftwNumbers")
+plan_transform{D}(::Fourier{D},x::Vector) =
+    error("transform for Fourier only implemented for fftwNumbers")
+plan_itransform{D}(::Fourier{D},x::Vector) =
+    error("transform for Fourier only implemented for fftwNumbers")
+
+transform{T<:Number,D}(S::Fourier{D},vals::Vector{T}) =
+    transform(S,vals,plan_transform(S,vals))
 
 function transform{T<:Number,D}(::Fourier{D},vals::Vector{T},plan)
     n = length(vals)
@@ -215,6 +223,9 @@ function transform{T<:Number,D}(::Fourier{D},vals::Vector{T},plan)
 
     cfs
 end
+
+itransform{T<:Number,D}(S::Fourier{D},vals::Vector{T}) =
+    itransform(S,vals,plan_itransform(S,vals))
 
 function itransform{T<:Number,D}(::Fourier{D},a::Vector{T},plan)
     n = length(a)
