+++ test/ChebyshevTest.jl
@@ -1,5 +1,9 @@
 using ApproxFun, Base.Test
+    import ApproxFun: testspace
 
+for d in (Interval(),Interval(1.,2.),Interval(1.0+im,2.0+2im))
+    testspace(Chebyshev(d))
+end
 
 ef = Fun(exp,Interval())
