+++ test/FourierTest.jl
@@ -1,6 +1,18 @@
 using ApproxFun, Base.Test
+    import ApproxFun: testspace
 
+for d in (PeriodicInterval(0.1,0.5),Circle(1.0+im,2.0))
+    testspace(CosSpace(d))
+    testspace(SinSpace(d))
 
+    testspace(Taylor(d))
+    testspace(Hardy{false}(d))
+
+    testspace(Laurent(d))
+
+    @show d
+    testspace(Fourier(d))
+end
 
 @test sum(Fun(1,CosSpace())) ≈ π
 @test sum(Fun([1],SinSpace())) == 0
